+++ packages/next-common/utils/consts/walletTypes.js
@@ -1,0 +1,8 @@
+const WalletTypes = {
+  METAMASK: "metamask",
+  POLKADOT_JS: "polkadot-js",
+  SUBWALLET_JS: "subwallet-js",
+  TALISMAN: "talisman",
+};
+
+export default WalletTypes;